+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fix getValue calls on Expressions in Laravel 10
+
 ## [1.1.0](https://github.com/clickbar/laravel-magellan/tree/1.1.0) - 2023-02-07
 
 ### Added